requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
Pillow>=10.0.0
pytesseract==0.3.10
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
numpy==1.24.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
openai>=1.0.0
gunicorn==21.2.0
